change Exchange names and Exchange subcomponent names for more semantic meaning and findability
