+++ app/components/Exchange_Processesing
@@ -2,7 +2,7 @@ import React from "react";
 import { clipboard } from "electron";
 import ReactTooltip from "react-tooltip";
 
-export default function Exchange_Process(props) {
+export default function Exchange_Processing(props) {
 	const { txData  } = props;
 	return (
 		<div>
