+++ app/containers/Dashboard.js
@@ -10,8 +10,6 @@ import FaArrowUpward from "react-icons/lib/fa/arrow-circle-up";
 import { NetworkSwitch } from "../components/NetworkSwitch";
 import WalletInfo from "../components/WalletInfo";
 import TransactionHistory from "../components/TransactionHistory";
-import Exchange from "../components/Exchange";
-import ShapeShift from "../components/ShapeShift";
 import SelectExchange from "../components/SelectExchange";
 import Support from "../components/Support";
 import Tokens from "../components/Tokens";
